Start getting user stats

+++ src/main/java/Core/Fight/FightUtil.java
@@ -7,7 +7,7 @@ public class FightUtil {
         return (Math.pow(lvl, 1.35) / 10) + 1;
     }
 
-    public static double calcEXPForLevel(int lvl) {
+    public static int calcEXPForLevel(int lvl) {
         // 5x^2 + 50x + 100
         return 5 * lvl * lvl + 50 * lvl + 100;
     }
